+++ backend/src/main/java/com/shallwecode/backend/user/application/service/UserService.java
@@ -1,6 +1,6 @@
 package com.shallwecode.backend.user.application.service;
 
-import com.shallwecode.backend.user.application.dto.FindUserListDTO;
+import com.shallwecode.backend.user.application.dto.user.FindUserListDTO;
 import com.shallwecode.backend.user.application.dto.user.FindUserDTO;
 import com.shallwecode.backend.user.application.dto.user.FindUserDetailDTO;
 import com.shallwecode.backend.user.application.dto.user.UserSaveDTO;
